Show the short hostname, not the FQDN in config_version
This is mainly a style and readability change.

Prior to this, on masters whose hostname is actually their FQDN, the
config_version script would show the entire FQDN. On nodes with really
longs FQDN's, it was not very nice to look at.

This takes the hostname of the master, splits it on dots (.) and takes
the first segment.

Now this: compile-master-02.int.lab.dmz.company-name.net-production-48fd18ab
Is this:  compile-master-02-production-48fd18ab